initial commit
this is a silly toy language supporting addition and multiplication of
ints with variable lookup. variables are introduced using let. All it
has right now is a type and an algebra. I want to use it to explore data
types a la carte, parsec, and how to do silly stuff with the jvm.